--- src/app/core/components/icons/Decor/Footer.js
@@ -1,12 +1,0 @@
-import React from "react"
-
-export default () => {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" viewBox="25 30 350 70">
-      <path
-        fillRule="evenodd"
-        d="M187.304 37.942c-.959 1.02-1.322 1.999-1.028 2.765.361.941-.157.799-2.507-.69-8.182-5.185-17.948-.823-20.856 9.316-.858 2.991.285 5.212 2.682 5.212 2.546 0 3.123-1.649 1.736-4.967-2.039-4.88 2.267-9.726 7.962-8.962 2.745.368 7.415 2.589 15.419 7.332l4.518 2.677-1.678 4.457c-2.385 6.337-1.717 7.409.976 1.566 1.982-4.301 2.443-4.858 3.7-4.482 6.348 1.899 12.489.452 16.933-3.992 5.168-5.168 4.554-11.187-1.191-11.679-3.61-.309-5.85 3.362-3.585 5.876 4.818 5.349-2.472 11.007-10.083 7.826-2.256-.942-2.331-1.085-1.917-3.671 1.244-7.779-6.328-13.643-11.081-8.584m6.98 2.23c2.321 2.321 2.842 7.639.749 7.639-1.579 0-5.804-3.454-6.211-5.078-.808-3.219 3.012-5.011 5.462-2.561m36.481 7.651c-6.439 1.932-5.386 6.151 2.217 8.884 5.12 1.841 5.677 2.816 2.238 3.919-5.125 1.642-6.994 3.814-10.563 12.27-.587 1.389-.782 2.525-.435 2.525.347 0 1.223 1.338 1.946 2.973 2.827 6.39 10.869 5.923 10.869-.632 0-2.363-.006-2.368-1.65-1.291-2.167 1.42-7.703-1.201-5.644-2.671.123-.088 1.652-.978 3.399-1.978 3.699-2.118 6.396-5.371 7.014-8.46.499-2.494 2.04-2.483 6.14.045 2.679 1.652 2.732 2.271.3 3.529-8.99 4.648-.697 16.066 11.62 15.996 11.802-.066 30.12-11.973 24.981-16.239-1.608-1.334-3.844-.198-3.709 1.885.218 3.368-.028 4.119-1.705 5.218-3.956 2.592-6.742-.645-8.686-10.089-1.898-9.223-5.079-12.522-12.078-12.526-5.074-.003-7.991 2.445-8.023 6.731-.001.186.207 1.512.463 2.948l.466 2.611-1.832-1.441c-1.007-.792-3.007-1.629-4.445-1.858-1.62-.259-3.105-1.043-3.906-2.061-.71-.903-2.699-2.717-4.42-4.031-2.807-2.144-3.1-2.621-2.862-4.665.147-1.252.251-2.251.233-2.219-.018.032-.886.314-1.928.627m-81.055 1.411c.492 1.883-5.298 7.857-9.067 9.354-1.243.493-2.666 1.209-3.162 1.591-.738.567-.82.14-.45-2.327.636-4.244-1.066-6.337-5.153-6.336-5.429.002-9.385 4.911-10.432 12.942-.828 6.348-6.968 7.285-6.968 1.064 0-2.381-2.32-3.305-3.113-1.24-3.947 10.287 24.812 16.981 29.08 6.769 1.386-3.319 1.23-4.227-1.132-6.589l-2.102-2.102 1.802-1.55c2.404-2.068 4.421-2.053 4.421.033 0 2.307 2.686 5.924 5.622 7.571 2.365 1.326 2.471 1.512 1.447 2.536-1.309 1.309-2.459 1.404-3.567.296-.628-.629-.808-.326-.808 1.361 0 2.735 2.399 4.144 5.021 2.95 2.423-1.104 5.654-8.217 3.733-8.217-.37 0-.673-.561-.673-1.246 0-2.107-3.834-6.743-6.052-7.317-3.019-.783-2.753-1.629.892-2.848 4.956-1.658 5.797-5.301 1.728-7.479-1.359-.727-1.446-.663-1.067.784m28.329 6.594c-11.262 6.868-6.967 27.493 7.313 35.12 9.559 5.106 24.076 2.584 24.076-4.182 0-3.698-5.092-4.237-5.816-.615-1.002 5.006-6.315 7.46-9.639 4.452-3.081-2.789-.481-6.503 7.303-10.432 9.575-4.834 12.844-10.97 9.513-17.852-3.053-6.306-10.31-7.098-15.695-1.713-3.711 3.711-4.522 4.061-4.522 1.954 0-6.205-7.122-10.031-12.533-6.732m-30.228 4.671c2.062 1.259 4.253 4.497 3.531 5.218-1.253 1.254-6.561-2.955-6.561-5.203 0-1.566.489-1.568 3.03-.015m89.283 5.414c-1.101 1.546-6.349 4.794-7.747 4.794-.849 0 1.177-4.534 2.769-6.196 4.041-4.217 8.155-3.059 4.978 1.402M39.876 63.851c-7.196 2.583-12.685 11.625-7.008 11.544 1.289-.019 2.201-.644 3.159-2.167 3.741-5.948 10.632-6.485 27.946-2.176 24.096 5.997 33.23 5.065 37.782-3.853 1.893-3.71 1.689-4.573-1.082-4.573-1.496 0-2.357.323-2.357.884 0 1.246-3.321 4.268-5.723 5.209-3.833 1.501-10.864.794-23.906-2.404-15.255-3.74-23.33-4.431-28.811-2.464m261.803-.641c-6.025 1.996-10.77 6.95-10.77 11.245 0 2.063 3.815 1.122 5.239-1.292 3.462-5.867 11.055-6.456 28.094-2.181 14.36 3.603 15.602 3.82 21.929 3.83 9.573.015 14.799-3.248 17.089-10.671.799-2.59-3.67-2.153-5.245.514-3.671 6.213-10.611 6.351-33.436.663-8.702-2.168-19.67-3.178-22.9-2.108"
-      />
-    </svg>
-  )
-}
