--- src/app/core/components/icons/Decor/Footer.test.js
@@ -1,11 +1,0 @@
-import "jest-styled-components"
-
-import { shallow } from "enzyme"
-import React from "react"
-
-import Footer from "./Footer"
-
-it("Render Footer, matches snapshot", () => {
-  const element = shallow(<Footer />)
-  expect(element).toMatchSnapshot()
-})
