+++ src/app/core/components/icons/group-beacons/Heart.js
@@ -1,0 +1,9 @@
+import React from "react"
+
+export default () => (
+  <svg viewBox="0 0 42 31" version="1.1" xmlns="http://www.w3.org/2000/svg">
+    <g stroke="none" stroke-width="1" fill="none" fill-rule="evenodd">
+      <path d="M20.84375,9.1953125 C21.7187544,6.44529875 23.0546785,4.35938211 24.8515625,2.9375 C26.6484465,1.51561789 28.8281122,0.8046875 31.390625,0.8046875 C34.4218902,0.8046875 36.8593658,1.73045949 38.703125,3.58203125 C40.5468842,5.43360301 41.46875,7.88279727 41.46875,10.9296875 C41.46875,13.6015759 40.8281314,15.7499919 39.546875,17.375 C38.2656186,19.0000081 35.8047057,20.5781173 32.1640625,22.109375 L31.6953125,22.3203125 C25.7265327,24.8359501 22.1093813,27.648422 20.84375,30.7578125 C19.5468685,27.6327969 15.9375296,24.820325 10.015625,22.3203125 C9.79687391,22.226562 9.62500062,22.1562502 9.5,22.109375 C5.84373172,20.5781173 3.37891262,19.0078205 2.10546875,17.3984375 C0.832024883,15.7890545 0.1953125,13.632826 0.1953125,10.9296875 C0.1953125,7.88279727 1.12108449,5.43360301 2.97265625,3.58203125 C4.82422801,1.73045949 7.25779742,0.8046875 10.2734375,0.8046875 C12.8515754,0.8046875 15.0351473,1.51561789 16.8242188,2.9375 C18.6132902,4.35938211 19.9531205,6.44529875 20.84375,9.1953125 Z" />
+    </g>
+  </svg>
+)
