+++ src/app/core/components/icons/group-social/index.test.js
